+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequest.java
@@ -1,14 +1,10 @@
 package org.apache.coyote.http11;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.UncheckedIOException;
 import java.util.HashMap;
 import java.util.Map;
 
 public class HttpRequest {
 
-    private static final String APPLICATION_X_WWW_FORM_URLENCODED = "application/x-www-form-urlencoded";
     private static final String JSESSION_ID = "JSESSIONID";
 
     private final HttpHeaders headers;
@@ -28,39 +24,6 @@ public class HttpRequest {
         this.requestURI = requestURI;
         this.protocol = protocol;
         this.body = body;
-    }
-
-    public static HttpRequest from(final BufferedReader bufferedReader) {
-        try {
-            final String requestLine = bufferedReader.readLine();
-            final var requestLineTokens = requestLine.split(" ");
-            final String method = requestLineTokens[0];
-            final String uri = requestLineTokens[1];
-            final String protocol = requestLineTokens[2];
-            final HttpHeaders headers = HttpHeaders.from(bufferedReader);
-            final Map<String, String> body = parseBody(headers, bufferedReader);
-
-            return new HttpRequest(headers, HttpMethod.of(method), HttpRequestURI.from(uri), protocol, body);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
-
-    private static Map<String, String> parseBody(final HttpHeaders headers, final BufferedReader bufferedReader)
-        throws IOException {
-        final boolean isURLEncoded = headers.containsHeaderNameAndValue(
-            HttpHeaderName.CONTENT_TYPE,
-            APPLICATION_X_WWW_FORM_URLENCODED
-        );
-        final int contentLength = headers.getContentLength();
-        if (contentLength > 0 && isURLEncoded) {
-            final char[] rowBody = new char[contentLength];
-            bufferedReader.read(rowBody, 0, contentLength);
-            final String body = new String(rowBody);
-
-            return X_WWW_Form_UrlEncodedDecoder.parse(body);
-        }
-        return new HashMap<>();
     }
 
     public boolean containsAccept(final String contentType) {
